[OPE] Send a heartbeat from the main process as well
